msdk: dec: reset async depth to one
Currently we use an async depth of 4 as default (based on recommendations
in msdk apps), which indicates how many asynchronous operations an application performs
before the application explicitly synchronizes the result. As a result, we
queue four frames in decoder which might not be good approach for
live streaming.

This patch reset the async-depth to 1 as default so that we do sync for
each frame we decode without queuing. Customer can play with already
exposed "async-depth" property for other use cases